cnn reorganising
